refactor(ShelterInfoCard): 재사용 컴포넌트를 위한 폴더 구조 변경
