+++ web/src/pages/homepage/components/MapView.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import { useEffect, useRef, useState } from 'react';
-import ShelterInfoCard from './ShelterInfoCard';
+import ShelterInfoCard from '../../../components/ShelterInfoCard';
 import theme from '../../../styles/theme';
 import { typography } from '../../../styles/typography';
